+++ usvm-fuzzer/src/main/kotlin/org/usvm/fuzzer/generator/primitives/DoubleGenerator.kt
@@ -6,9 +6,10 @@ import org.usvm.fuzzer.generator.GeneratorContext
 import org.usvm.fuzzer.generator.GeneratorSettings
 import org.usvm.fuzzer.util.UTestValueRepresentation
 import org.usvm.instrumentation.testcase.api.UTestDoubleExpression
+import org.usvm.fuzzer.generator.random.nextDouble
 
 class DoubleGenerator(): Generator() {
-    override val generationFun: GeneratorContext.() -> UTestValueRepresentation = {
+    override val generationFun: GeneratorContext.(Int) -> UTestValueRepresentation = {
         val randomDouble = random.nextDouble(GeneratorSettings.minDouble, GeneratorSettings.maxDouble)
         UTestValueRepresentation(UTestDoubleExpression(randomDouble, jcClasspath.double))
     }
